backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
django-cors-headers>=4.2.0